ex7

+++ src/utility/Element.java
@@ -1,5 +1,5 @@
 package utility;
 
 public abstract class Element implements Comparable<Element>, Validatable {
-    abstract public Long getId();
+    abstract public Integer getId();
 }
